fix(market): auto-open the trading market on boot in every run profile
The market only auto-opened in dev (DevMarketOpener + MarketSelfHealScheduler were @Profile("dev"); ScenarioRunner opens it only while seeding). So a non-dev run with seed.mode=off (the persistent Supabase / jpa profile) came up READY but never OPEN, and every reserve threw MarketNotOpenException.

Rename DevMarketOpener -> MarketOpener (Infrastructure/scheduling) and broaden both it and MarketSelfHealScheduler to @Profile("!test"): the market opens at boot and self-heals in every real run (dev/jpa/supabase/prod). Admin UC-32 open/close is unchanged and ensureMarketOpen() still respects a deliberate close; tests (@Profile("test")) are unaffected.
