+++ src/main/java/com/ticketing/system/Infrastructure/scheduling/MarketSelfHealScheduler.java
@@ -9,16 +9,16 @@ import com.ticketing.system.Core.Application.services.SystemAdminService;
 import lombok.extern.slf4j.Slf4j;
 
 /**
- * Self-heals the trading market in the {@code dev} profile (#455). If a transient external-service
- * outage — e.g. a cold-starting WSEP endpoint — left the market closed at boot, this periodically
- * re-attempts the open so the system recovers <i>without a restart</i> (V3 Req 6).
+ * Self-heals the trading market in every real run (#455). If a transient external-service outage —
+ * e.g. a cold-starting WSEP endpoint — left the market closed at boot, this periodically re-attempts
+ * the open so the system recovers <i>without a restart</i> (V3 Req 6).
  *
  * <p>{@link SystemAdminService#ensureMarketOpen()} is idempotent and respects an admin's deliberate
- * close, so this is a no-op once the market is open (or was closed on purpose). Dev-only: production
- * opens the market through an admin (UC-32), which the WSEP handshake retry already makes reliable.
+ * close, so this is a no-op once the market is open (or was closed on purpose). Runs in every profile
+ * except {@code test} (tests drive the market state themselves); an admin can still open/close via UC-32.
  */
 @Component
-@Profile("dev")
+@Profile("!test")
 @Slf4j
 public class MarketSelfHealScheduler {
